+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- The next version number should be the version from pubspec.yaml - 5.2.0 -->
 ### Fixed
 - Correct filtering of sport events [#3986](https://github.com/rokwire/illinois-app/issues/3986).
+### Changed
+- Replaced the depreciated widget WillPopScope with new PopScope widget [#3832](https://github.com/rokwire/illinois-app/issues/3832).
 
 ## [5.1.32] - 2024-02-19
 ### Removed